pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager"
authors = ["choolkov <s.v.chulkov@ya.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.2"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
dj-database-url = "^1.0.0"
django-bootstrap4 = "^22.2"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.10.0"
wemake-python-styleguide = "^0.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
